requirements.txt
absl-py==1.4.0
astunparse==1.6.3
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
contourpy==1.0.7
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
deprecation==2.1.0
filelock==3.10.7
Flask==2.2.3
flatbuffers==2.0.7
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
html5lib==1.1
huggingface-hub==0.13.3
idna==3.4
img2pdf==0.4.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
langdetect==1.0.9
Levenshtein==0.20.9
libclang==16.0.0
lit==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.0.4
mplcursors==0.5.2
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.12.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pdf2jpg==1.1
pikepdf==7.1.2
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pydyf==0.6.0
PyMuPDF==1.21.1
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==3.21.1
pyphen==0.14.0
python-dateutil==2.8.2
python-doctr==0.6.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.13.7
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
Shapely==1.8.5.post1
six==1.16.0
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tf2onnx==1.14.0
tinycss2==1.2.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typeguard==3.0.2
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
zopfli==0.2.2
